+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.17.2 (2026-07-27)
+
+
+### EZ Policy Enforcer
+
+* **enforcer:** defer isolate boot loop to fix startup race condition
+
 ## 0.17.1 (2026-07-20)